// src/common/types/enums.ts
export enum EAPI_KEY {
  CITY = 'city',
  CACHE = 'cache',
}

export enum EAPI_STATUS {
  IDLE = 'idle',
  LOADING = 'loading',
  SUCCEEDED = 'succeeded',
  FAILED = 'failed',
}

export enum EMETRIC_TYPE {
  TEMPERATURE = 'temperature',
  PRESSURE = 'pressure',
  WIND_SPEED = 'wind speed',
  HUMIDITY = 'humidity',
}